readyset-psql: Rename `Value` to `TypedDfValue`
The named `Value` is very overloaded in the psql code, and doesn't do
enough to explain what it actually represents. Renaming to
`TypedDfValue` should help avoid mixing this type up with other types
and type parameters named `Value`, and also makes it a lot more obvious
that this type is just a wrapper for a `Type` and a `DfValue`.

Change-Id: Ic7064576b7ed398480d705c95780a2ec17548390
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5491
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Reviewed-by: Jason Brown <jason.b@readyset.io>