migrate: Decouple DomainMigrationPlan from DfState
Rather than passing in a *full* DfState (which is large, and complex,
and difficult to construct) to DomainMigrationPlan:new, just pass in
the bit of DfState we actually care about. This paves the way for
constructing a DomainMigrationPlan more easily in unit tests, which is
useful because it allows us to unit test things like the materialization
planner.

Change-Id: Ie0df6d1ca6b01c380403132a596d81a84b0b6e9a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5494
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Reviewed-by: Fran Noriega <fran@readyset.io>
Tested-by: Buildkite CI