psql: Rename psql-srv `Value` enum to `PsqlValue`
The name `Value` is still very overloaded, and doesn't give much context
as to what type of value is being referred to. Similar to how DfValue is
used in many crates and so has the prefix "Df" to help disambiguate, I'm
adding the prefix "Psql" to make psql-srv's `Value` type name clearer.

Change-Id: Ifd370fe7117fbb36def5006bf411d7e3e2bad3e2
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5492
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Tested-by: Buildkite CI