dataflow: Allow performing full replays to only some replicas
Add a `replicas: Option<Vec<usize>>` field to
DomainRequest::StartReplay, which if set will set the same new field on
ReplayPieceContext::Full, and cause all Fanout senders to send the full
replay to only those replicas. Currently this is always set to None, but
will be set to a value in the future to allow replaying to only some
replicas of downstream domains (eg, those that are being newly
scheduled).

Change-Id: I243ad0b7fb8ae77e167f1715519cee03481d9ccc
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5495
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>