psql: Simplify lots of code by eliminating TryInto bound
We inevitably convert values into PsqlValue anyway when we encode them,
so rather than allowing implementors of PsqlBackend to use any type that
implements TryInto<PsqlValue>, it turns out to simplify things a lot if
we just demand that the implementations yield PsqlValue directly.

In some cases, the code is already pretty much doing this anyway, and in
cases where it isn't, it's much simpler to just add an extra line of
code to convert to PsqlValue than it is to carry around all the weight
of all the extra generics everywhere else in the psql code.

This also will facilitate the upcoming change of eliminating the
intermediate TypedDfValue conversion step in Resultset.

Change-Id: I9b450c4f83cce81922b011576cd742209657dc88
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5501
Reviewed-by: Jason Brown <jason.b@readyset.io>
Reviewed-by: Aspen Smith <aspen@readyset.io>
Tested-by: Buildkite CI