mysql-srv: Fix COM_STMT_EXECUTE packet data read error
COM_STMT_EXECUTE packet has a one byte flag to indicate if the
parameters need rebinding (new_params_bind_flag). When the same
prepared statement is executed twice, the flag is zero.

The code was handling the content when the flag was enabled but
if the flag was set to zero, it was not removed from input stream.
This made all the parameter values bit-shift by 8.

This fix removes the flag byte if it's set to zero.

Change-Id: I9525c00c5f0f206ecdf3f7b9e0cddf56b645b3cd
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5503
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>