materialization: Simplify condition for full mat idempotency
In order to make sure we don't double-replay to the same full
materialization when we're only adding an index to it, we added a
condition in 17b5bee8a (Make the materialization planner (more)
idempotent, 2021-04-07) that avoids generating paths and pending replays
for nodes if we already have replay paths for that node. This commit
slightly simplifies the way that works, by storing a simple boolean
indicating that we already have paths for the node in `Plan` rather than
a full clone of the map of paths - this also requires tweaking an
`unwrap()` after plan construction in Materializations itself.

Change-Id: Icd87ccb0c2f55f586f02d993c89745c00367e472
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5516
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>