migrate: Make pending recovery a setting on DomainMigrationPlan
Rather than having pending recovery be a field of Materializations,
initialized to false during deserialization and set to true after the
first time `commit` is called, make it an (enum) field on
DomainMigrationPlan set each time we run a migration, with variants for
"recover" and "extend". Other than that, the field is used the same way.

Refs: REA-3138
Change-Id: Ie07eacce9cb3492553f4d73c318a9d4eb1b17f5c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5517
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>