replication-offset: Add crate
This commit introduces a new replication-offset crate that houses all of
the types and conversion code pertaining to the representation of
replication offsets for MySQL and Postgres. Previously, the
`ReplicationOffset` type was defined in readyset-client, and the types
wrapping the actual Postgres/MySQL replication offset values
(`PostgresPosition`/`BinlogPosition`) lived in replicators. Moving all
of these types into the same crate sets the stage for redefining
`ReplicationOffset` to be an enum that simply wraps `PostgresPosition`
and `BinlogPosition` (this was previously not possible because it would
have introduced a circular dependency between replicators and
readyset-client). This change will also allows us to deduplicate the
`Display` implementations for these types.

This commit also renames `BinlogPosition` to `MySqlPosition` to be
consistent with `PostgresPosition`.

Refs: REA-3118

Change-Id: I39818a092b177871cdc178736c2a18086a7fca56
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5604
Tested-by: Buildkite CI
Reviewed-by: Aspen Smith <aspen@readyset.io>