dataflow: Track completed full replays per-node
Previously, there was a field on `domain` called `replay_completed`,
which tracked whether that domain (or any node in that domain, I...
guess?) had received a complete full replay. That field was initialized
to `false`, then *any* node within the domain receiving a full replay
would set it to `true`. Then, the `QueryReplayDone` message would query
the value of that field and then... set it to `false` again. I'm not
entirely *sure* why that's happening, but @eeeeeta says:

> because you could have another replay??

Which I suppose I sort of understand, but this causes issues, primarily
by being highly non-idempotent - specifically, when querying whether
replays are done to multiple replicas of the same domain, if one has
completed but not the other we will *keep querying* - meaning that
because `replay_completed` had been set to false on the one that *did*
complete, we can never actually finish querying for replays to finish on
all replicas even after the other one finishes.

To make things properly idempotent and also a *lot* more precise and
granular, this commit changes the tracking of "has this state received a
full replay" to be per-node (on the State for internal nodes and on
ReaderWriteHandle for reader nodes), and the QueryReplayDone message to
also specify a (local) node index that they want to query for.

It's separately still a bit of an efficiency issue that we're always
querying all replicas for whether they've received full replays rather
than only those which haven't yet - but that can be fixed elsewhere.

Refs: REA-3138
Change-Id: I0d162971fdc6fe4a7f6b9e25189f983f215f536a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5518
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>