nom-sql: Add parsing support for ROW expressions
Add parsing support for postgresql ROW expressions[0], which can be
either the keyword "ROW" followed by a comma-separated parenthesized
list of exprs, or a bare comma-separated parenthesized list of two or
more exprs. This commit only allows parsing them, without adding support
for evaluating them or representing them in dataflow at all.

I'm doing this now because I'm going to use row exprs as a trick to fix
the precedence of parsing IN and NOT IN - I don't intend to add eval
support for this any time soon.

Change-Id: I53c005f8e13de118ba98af8606d1a1a2d8a1c360
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5520
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>