nom-sql: Fix precedence of (NOT) IN
Fix the precedence of the IN and NOT IN operators by having them
participate in the overall operator-precedence Pratt parser rather than
parsing them via an expr parser directly. Since their rhs is a special
AST node rather than Expr itself, this requires playing some tricks with
`Expr::Row` on the RHS. This leaves one edge case at the moment - the
following exprs parse the same:

    lhs IN ((rhs1, rhs2))

and:

    lhs IN (rhs1, rhs2)

but the former *should* parse the same as:

    lhs IN (ROW(rhs1, rhs2))

Since we don't currently *support* ROW exprs within dataflow, I'm
leaving this unfixed with an ignored test for now, but we should fix it
if we ever add support for those.

Fixes: REA-3163
Change-Id: I72b683dad29df0854672e64edd080a14324bb305
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5521
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>