migrate: Reuse tags for existing paths
We're now going to be running the materialization planner in "recovery
mode" multiple times for the same set of nodes, specifically if we have
to set up a new replica of a domain for which replicas are already
running. When we do this, we don't want to create duplicate tags for
paths we've already planned - not really for *correctness* reasons, but
primarily because that's an inefficiency and source of confusion that
can grow arbitrarily large as workers die and re-join the cluster.

This commit makes sure that if we make a replay path we already have, we
use the same Tag we've already generated for it. To achieve this, we add
the Index for the path to the right-hand side of
`Materializations::paths`, and convert it to a `BiHashMap` rather than a
`HashMap`, so we can look-up based on the right-hand side (the list of
nodes in the path and, now, the index. Both of these are necessary to
uniquely identify the path, since a single index can have multiple paths
due to diamond unions, and there can be multiple paths that follow the
same nodes to different indexes). Then, when making tags for paths in
Plan, we try to look up our pathh in that map before generating a new
tag.

Change-Id: I36c7205751304f108df768e3f50fbc9cc403c72c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5528
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>