psql-srv: Implement FromSql for PsqlValue
This is not used yet, but the intent is for this to take the place of
the current code path that is used for the proxy code where we call
DfValue::from_sql followed by PsqlValue::try_from (with an intermediate
step of wrapping in TypedDfValue). As such it's intended to basically be
a composition of those two functions, but avoiding the intermediate
DfValue step lets us save a lot of redundant encode/decode work.

Refs: REA-3143
Change-Id: If465b651034a0a6f93ee6ab10686cccfa283718a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5538
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>
Reviewed-by: Aspen Smith <aspen@readyset.io>