server: Only target full replays at just-placed replicas
If we've run migration to recover a new replica of an already-running
domain that contains full materializations, we only need to send full
replays for those materializations from upstream nodes to the replicas
that are running now, but weren't running before. This commit does that,
by first tracking the eventual target domain of each full replay (on the
StartReplay request, since there's nowhere else to put it) and also
tracking a map of the replicas that were placed during
DomainMigrationPlan::apply. Then we look up the target domain in that
latter map and use the list of replicas (optionally for the shard we're
targeting, if we're only targeting one shard) to set the `replicas`
field on StartReplay before sending it to the source domain.

Change-Id: Ie4f0e678b7006efea44acb26a4ee18400c458b1b
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5536
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>