readyset-psql: Switch proxy path to use new PsqlValue FromSql code
This should improve proxy efficiency (measurements forthcoming), and
also paves the way for eliminating the last of the unnecessary DfValue
enum index decoding that keeps us from fixing REA-3143.

This diff is a little bit misleading, because the actual change happens
in the ResultsetInner::Stream arm of the match; by moving the conversion
code so that it only applies to the ReadySet arm, the Stream arm infers
that we need Vec<PsqlValue> instead of Vec<DfValue> and switches the
try_get call to use our new FromSql impl on PsqlValue.

Refs: REA-3143
Change-Id: Ie9cc0f7b9124728a5db51de3e20f4580886e19d7
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5539
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>
Reviewed-by: Aspen Smith <aspen@readyset.io>