readyset-data: Don't decode enum indices in DfValue's FromSql impl
Now that we're no longer using this code for proxying, we can change it
to match the other Postgres code, which fixes REA-3143.

Other than the aforementioned bug, it didn't hurt anything to decode
them here, but it is now redundant since we already decode them in all
the other necessary places for replication/writes/query parameters/etc.

Fixes: REA-3143
Change-Id: I8da9a33669ccbab09e3e0a9013c12d040068fe69
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5540
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>
Reviewed-by: Aspen Smith <aspen@readyset.io>