server: Don't fail node_sizes if some replicas are unplaced
Convert `node_sizes` from using `query_domains`, which internally fails
if some replicas are unplaced, to manually calling `send_to_healthy`,
which returns None for unplaced replicas, and then flattening out those
Nones so that we count unplaced replicas as having no state in their
nodes (which is correct). This allows running graphviz on a deployment
when not all replicas are scheduled.

Fixes: REA-3141
Change-Id: I553d822e2eb0062c14e2ab0196d1df461c605d89
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5541
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>