Adapter: Remove check_server_version_compatibility()
Removes check_server_verion_compatibility() which is used to ensure that
the adapter and server have compatible versions. However, this check was
blocking and prevented the adapter from accepting new connections. We
would like to allow the adapter to run before a server has joined the
cluster.

This check should instead be run every time the adapter is informed of a
new controller.

Ref: REA-3176
Change-Id: I18c053621947aa7cd98c3a26ed4557e053369276
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5548
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>