server: Don't send QueryReplayDone to completed replicas
As an optimization, avoid sending QueryReplayDone messages to replicas
that have already completed replays (or will not receive replays,
because they were already running before the DomainMigrationPlan was
applied). This is done by keeping track of a BTreeSet of the replicas
we're waiting for, and passing that to a new `send_to_healthy_replicas`
method on `DominHandle` to only send to those replicas.

Fixes: REA-3174
Change-Id: I184d47d61f7d0e0d9c77ff12e00a43da0f8b639c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5549
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>