replication-offset: Make repl offset an enum
Previously, `ReplicationOffset` was defined as a struct that was
overloaded with domain knowledge about the representations of both the
MySQL and Postgres replication log positions. Building upon previous
commits, this commit redefines `ReplicationOffset` to be an enum that
wraps `MySqlPosition` and `PostgresPosition`, delegating conversion,
comparison, and display to the wrapped types.

Refs: REA-3118

Change-Id: I680e22272d91d85335f84411c3315772f93c6b78
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5607
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>