adapter: Create caches for aliased queries
Make ViewsSynchronizer explicitly cache queries that it detects have
been migrated, so that new aliases for existing caches can be queried.

In the case that the query was cached by a different adapter, this is a
no-op in the server.

In the case that the query was an alias of another query that got
cached, an entry will be made for the former in the ExprRegistry alias
map, which enables ReadySet to serve it.

Change-Id: I524125d8d73b0f8238d2e79551800e940ac76522
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5552
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>