server: Always add indexes during recovery
We might run `Materializations::extend` in recovery mode for a set of
nodes which already have indexes in Materializations, eg if we are
recovering a domain for which we're already running replicas or if we're
recovering a domain that was previously running on a worker that failed.
In this case, we still need to set up the indexes and (existing!) replay
paths for the nodes within that domain. This does that, by passing a
reference to `DomainMigrationPlan` into
`Materializations::extend` and checking if it's in recovery mode in a
few conditions.

Refs: REA-3002
Change-Id: I53c21f809801843834d199d6bdfedbd7ef10b3db
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5566
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>
Reviewed-by: Dan Wilbanks <dan@readyset.io>