Adapter: Query stats persisted in authority from adapter
Adds an Arc<Authority> field on Backend to allow querying persisted
stats from the adapter. The field was added rather than exposing the
existing authority client in NoriaConnector, since that client was not
designed to be used outside the context of querying for the current
Controller address.

The Backend will now query and return the persistent status as part of
the SHOW READYSET STATUS response.

Change-Id: Ibd1368282f41fb8d5256fe1885d764db3ce7f4af
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5569
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>