server: Remove conns for replicas that die
Within a running Replica, we maintain a map of established connections
for other domain replicas, to avoid having to re-establish a connection
on every message. If one of those replicas *died*, however, we'd never
remove the established connection from that map, meaning we'd never be
able to send messages to it if it ever came back.

Now, the `ChannelCoordinator` has a `tokio::sync::broadcast` channel
which can notify interested subscribers when the socket address for a
ReplicaAddress is changed - we subscribe to this in the Replica's
select-loop to remove the entry from that map of connections when we're
removing the configured socket from the channel coordinator.

This fixes an issue where after a worker died and came back, it'd never
be able to receive domain messages (such as full replay packets!).

Refs: REA-3002
Change-Id: If4defb00e9dd192a1def76bb1034723027f00130
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5578
Reviewed-by: Luke Osborne <luke@readyset.io>
Tested-by: Buildkite CI