vertical: Add test for "incongruent" joins
This specifically exercises an eviction edge case described in the Noria
thesis, where we have to be careful to propagate join evictions based on
the correct key due to the query key differing from the join key.

I wrote this expecting that it might reveal an eviction bug, but it
seems to be passing locally for me. Still seems worth adding to the
suite though, both for guarding against future regressions, and for
exercising the query more heavily in the nightly runs.

Refs: REA-3029
Change-Id: I29fe0264662e246265f7482483542a87122bb71a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5573
Reviewed-by: Fran Noriega <fran@readyset.io>
Tested-by: Buildkite CI