vertical: Delete old dataflow vertical test regressions file
All of these are old enough that the test seeds almost certainly don't
reproduce the same test cases anymore (this is one of the gotchas with
the proptest regression persistence - if the test changes enough, the
same seed may no longer generate the same test case it used to).

They're also from before shrinking was added to the vertical dataflow
tests, so the commented minimal failing cases aren't very illustrative
of anything either, since the failing cases are too complex to clearly
see what the bug was at a glance.

I think it's better to clear this file out, so that we're not wasting
time running through all the same seeds in the regressions file every
time we run these tests.

Change-Id: Ic035bfd44155a2ad7d2a812d9d66f45da74b45ac
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5575
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>