controller: Drop down domain request backoff significantly
Previously, due to a misunderstanding of the options for
ExponentialBackoff, we were using a *base* of 200ms as the domain
request backoff, which meant that the request times were 200
milliseconds, then 40 seconds, then 2.2 hours (!!). This creates all
sorts of problems with extremely long request timeouts, including an
extend_recipe timeout that I saw in a clustertest. This changes that to
use a base of 2 milliseconds instead (which since it's exponential
really ought not to over-request), with a constant multiplication factor
of 2 just to keep the requests limited even a little bit.

Change-Id: I487e3e728ba35d0a6ea0034a10953d42e282d3d4
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5576
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Reviewed-by: Fran Noriega <fran@readyset.io>