server: Wait for the right nodes to be ready
f5e9b05b1 (migrate: Wait for IsReady only as we need to replay,
2023-05-18) introduced the `non_ready_nodes` set passed around during
materialization planning to allow waiting to check if a base table had
initialized until we knew it needed to be replayed from, but had a
mistake - it was tracking the node that was the *target* of the replay,
not the node that was the *source* of the replay. This was obscured in
tests by the fact that the IsReady request timeout is long enough that
if we initialized two base tables at once, it was likely that the second
would finish initializing before the first - but upon lowering this
timeout, we hit a scenario where we could try to request a replay from a
non-ready node. This fixes this by making the `non_ready_nodes` set
contain (DomainIndex, LocalNodeIndex) tuples instead (since that's the
way we refer to nodes in PendingReplay) and using the *source* node of
the replay as the "non-ready" check in the set.

Change-Id: Ib32540a5ff5b2c276e45470ccd9201bae4eed43f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5596
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>