replicators: Fix XLogData LSN usage
Postgres's XLogData message in the streaming replication protocol
contains two separate LSNs: one that is associated with the start of the
data in the message and one that is associated with the end of the WAL
on the server. Previously, we were using the latter as the LSN
associated with the event contained with XLogData messages. This is not
correct: though these two LSN values may be equal if we haven't fallen
behind in replicating, as soon as there is replication lag, the LSN
associated with the current WAL event could be less than the LSN of the
end of the WAL on the server.

This commit addresses this discrepancy by using the LSN in the XLogData
message that is associated with the start of the WAL data in the
message. Additionally, keepalive messages sent to us from Postgres only
contain the LSN of the WAL end on the server, which makes sense because
keepalive requests don't pertain to any data mutation and thus would
have no LSN of their own. To account for this, this commit also makes it
possible for `WalEvent`s returned by the WAL reader not to have an LSN
at all.

Refs: REA-3118
Release-Note-Core: Fixed an issue with Postgres replication that could
  cause dropped replication events if ReadySet fell behind while
  replicating

Change-Id: I3e460d0e99b7a7efd4a4b2baaf2bf5710b76fb5e
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5603
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>
Reviewed-by: Dan Wilbanks <dan@readyset.io>