Replicators: Surface more stats from replicators
Expands the existing replicator -> Controller mpsc error channel to
handle a new `ReplicatorMessage` enum. This expanded channel replaces
the Notify channel, since it now handles events in addition to errors.

The new channel passes a notification to the Controller when replication
starts, which is added to a new
`PersistentStats::last_started_replication`, and surfaced as part of the
ReadySet status message.

Change-Id: Ibe209da6ffb769d9a1070bc1e582fa16db1d5e4a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5600
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>
Reviewed-by: Ethan Donowitz <ethan@readyset.io>