readyset(-server): Remove noria_metrics/CompositeMetricsRecorder
Previously we supported two different metrics recorders, though one
(--noria-metrics) was hidden from the default confiuration options.

We haven't used --noria-metrics for a while, so this commit removes it
as well as the CompositeMetricsRecorder, which allowed both it and the
currently-used prometheus-metrics to be enabled.

We are left with just prometheus.

Change-Id: I90ddf68ce6c8b277396dcf894b7b2780b0b0fbd0