replication-offset: Redefine `MySqlPosition`
Previously, `MySqlPosition` was defined only as the raw binlog file name
and the position within that file. The simplicity of the definition here
didn't matter much here because we converted the information in the
file name and position into an orderable 128-bit integer when converting
from `MySqlPosition` to `ReplicationOffset`. The downside to that
approach is that the implementation of `ReplicationOffset` has to rely
on MySQL-specific domain knowledge in order to properly implement
`Display` for `ReplicationOffset`; specifically, `ReplicationOffset` needs
to know implementation details about how to decode the MySQL-specific
information from the 128-bit integer stored within `ReplicationOffset`.
This same conversion knowledge also needs to live in the implementation
of `MySqlPosition` in order to implement the conversion from
`MySqlPosition` to `ReplicationOffset` and to implement `PartialOrd`.

Having this domain knowledge exist in multiple places leads to code
duplication and error-prone conversions involving trying to fit all of the
relevant `MySqlPosition` information into a 128-bit integer. A better
approach here would be to redefine `ReplicationOffset` to be an enum
that simply wraps `MySqlPosition` and `PostgresPosition`, limiting the
existence of domain knowledge about internal representation and ordering
to the database backend-specific types themselves.

This commit lays much of the groundwork for that change by redefining
`MySqlPosition` to consist of four parts: the base name of the binlog
file, the suffix of the binlog file (which is used to order files within
a given binlog stream), the length of the suffix (which is needed
because suffixes are padded with zeroes in the file names), and the
integral position within the file. This allows us to easily define
`PartialOrd` for `MySqlPosition` directly instead of having it defined
on `ReplicationOffset`. In a future commit, when `ReplicationOffset` is
redefined as an enum wrapper, `ReplicationOffset`'s `PartialOrd` and
`Display` implementations will simply delegate to the types it wraps,
ensuring that it needs no access to any domain knowledge of the
underlying database backends.

Refs: REA-3118

Change-Id: I31469da3268656175ab67dc2130de36942f11491
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5605
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>