readyset-data: Fix TimestampTz error for BC leap days
The old version of the code failed for the reasons described in the code
comments. This commit fixes the implementation, and adds a test for the
corner case that prompted this fix.

Note that the previous version of the code had some special casing to
add a + or - prefix if needed, in order to make sure we handle years
with five or more digits. But since this match arm specifically handles
BC years, we don't need to worry about needing a + prefix, and
formatting a negative integer as the year in the "fixed" string will
automatically add a - sign if needed, so I removed the old code for
adding prefixes as part of this change.

Release-Note-Core: Fix a bug that led to table replication failures on
  Postgres when we tried to replicate temporal data that contained leap
  days in years prior to the year 1 CE.
Fixes: REA-3193
Change-Id: Id030c60f3260ee51ce5050f1776a3b1e431899f3
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5601
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>