integration: Add failing test to reproduce REA-3207
I found this via a join query that I tried adding to the vertical
dataflow tests. It currently fails with this error:
```
thread 'Domain 1.0.0' panicked at 'assertion failed: `(left == right)`
  left: `[Int(2), Int(99)]`,
 right: `[Int(1), Int(99)]`', dataflow-state/src/keyed_state.rs:247:17
```
The failure specifically only seems to occur if we use the same foreign
join key on both rows.

Note that this bug does not seem to be backend-specific, as this same
test can easily be ported to the readyset-psql integration test suite
and the same error occurs.

Refs: REA-3207
Change-Id: Ia6000191cb811287eb27df4fc794eafaa6f529c5
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5617
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>