ddl_vertical: Increase number of operations per test case
The newly added Evict op is very quick to execute, and we should
probably start generating more ops per test case anyway to increase the
odds of different sets of operations occurring in the same test case.

The current 1,000 test case run that we do in nightly only takes around
two hours last I checked, so this should still finish in plenty of time
without becoming a bottleneck and slowing down the build as a whole.

Refs: REA-3203
Change-Id: I292f0c60f691fffbc4f7d169af50dd08bfef70e4
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5619
Tested-by: Buildkite CI
Reviewed-by: Aspen Smith <aspen@readyset.io>
Reviewed-by: Dan Wilbanks <dan@readyset.io>