errors: Remove "invariant" from the description of Internal
This word is nice when invariant!/invariant_eq!/invariant_ne! are used
to assert boolean invariants, but is not especially useful to users in
the general case for `internal!` and `internal_err!`. This commit
removes it from the general Internal error message description and moves
it to the individual descriptions on the boolean `invariant`
macros (instead of "assertion", since that's really more for panics).

Change-Id: Ided737f5ce05209d8fe58fe8f9f65ac421cef602
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5620
Reviewed-by: Ethan Donowitz <ethan@readyset.io>
Reviewed-by: Nick Marino <nick@readyset.io>
Tested-by: Buildkite CI