server: Add a WorkerRequestKind to kill domains
When domains within a multi-instance deployment die and then come back,
we need to restart any downstream domains in order to ensure their state
is correct. This commit starts the work required to do that, by
introducing a new `WorkerRequestKind::KillDomains` worker request, which
kills a list of domains running on a worker. Currently nothing sends
this yet.

Change-Id: If936814f386da660d867fec400af415dfbaa8ad3
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5624
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>