server: Kill domains downstream of failed workers
When the controller is notified that some workers have died, collect a
list of all the domains *downstream* of the ones running on those
workers, and send requests (in parallel) to the workers running those
domains to kill them.

We need to do this since if an upstream domain has died, we have no
guarantee that the state stored within the downstream domain remains
correct (since any messages for state updates might have been dropped).

There's a subtle bug here that will need to be fixed in a follow-up
commit - if the worker that *was* running one of those downstream
domains rejoins but the upstream domain remains unschedulable (eg
because it's got a base table whose volume isn't available in any
running worker), we still run the downstream domain! We need to instead
make it so that we never run a domain if we haven't been able to run its
upstream domains. This'll be fixed later, but is less of an issue for
now since we're only ever running with one server (so all domains
upstream of readers and the controller itself will be running on the
same instance) - in fact, the only reason I'm implementing this right
now is because it paves the way for granular handling of domain failure.

Refs: REA-2406
Change-Id: Ic392a51749470894527835b8f9585cd22a216251
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5625
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Tested-by: Buildkite CI