replicators: Handle out-of-order PGSQL LSNs
Previously, we assumed that the LSNs we received from replication events
from Postgres increased monotonically. In reality, this is not the case.
Postgres reorders replication events on a per-transaction basis so that
replication log consumers receive the events in a given transaction all
at once. Since transactions can happen concurrently, this means that we
are not receiving events in the order in which they were written to the
WAL, which means the LSNs will not increase monotonically. Instead, it
is the case that the LSNs associated with *commits* will increase
monotonically, and the the LSNs of events within each commit will also
increase monotonically.

We relied upon the monotonicity of LSNs in order to guarantee the
consistency of data across base tables and after replicator
restarts/resnapshots. To account for our more nuanced understanding of
LSN ordering, this commit redefines `PostgresPosition` to consist of an
ordered pair of the LSN of the last commit we handled and the LSN of the
current WAL event. Since LSNs of commits and LSNs of events within
transactions increase monotonically, this new defintion gives us a strong
ordering that we can use to ensure the consistency of the data we
replicate across different base tables.

Release-Note-Core: Fixed an issue where we were incorrectly assuming that
  LSNs in replication events from Postgres would always increase
  monotonically
Fixes: REA-3118, REA-3142

Change-Id: I1996748c49970c2c1cb30b1dad978e08cbe118e8
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5608
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Reviewed-by: Luke Osborne <luke@readyset.io>