client: Allow making controller handle from raw URL
Sometimes instead of having an authority handle, we have access to the
raw URL of the controller itself. This is currently the case inside the
Worker, which gets notified *proactively* about changes to the URL of
the controller. We'd like to be able to make *requests* to the
controller in this case using the same mechanism and wrapper methods we
use to make requests to the controller via the ReadySetHandle.

To that end, this commit introduces some new functionality to the
readyset_client::controller module: First, a RawController struct, which
wraps the controller Url and some extra state and config and impls
Service<ControllerRequest> in the same way Controller does albeit
without the retry logic used to re-look-up the controller's URL (using a
factored-out `controller_request` function which performs the actual
logic for both). Then, the inner service inside of ReadySetHandle is
replaced with a `tower::util::Either` of Controller and RawController,
and a new `make_raw` constructor method is added, so that we can have a
"raw" version of the ControllerHandle which can be constructed from only
a URL and used to make the same set of controller requests.

Change-Id: If6097416af640158ceb62802e7c62d8a299c3c4b